feat: refactor background component
